Refresh data (#29)
* Add manual data refresh functionality to UI

Introduces a 'Refresh Data' button to the UI, allowing users to manually trigger data reloads. Implements a reactive trigger to coordinate data refresh for data tab and updates the experiment accession observer to respond to this trigger.

* Update organization diagram on landing page

Replaces the old organization diagram with a revised image in the landing page UI. Adds the new 'research_ISPI_organization_revised.png' to the project.